fix and split deploy-suite script
