+++ package.json
@@ -19,6 +19,7 @@
   },
   "devDependencies": {
     "@appliedblockchain/silentdatarollup-core": "^0.1.3",
+    "@appliedblockchain/silentdatarollup-ethers-provider": "^0.1.3",
     "@appliedblockchain/silentdatarollup-hardhat-plugin": "^0.1.3",
     "@nomicfoundation/hardhat-chai-matchers": "^2.0.0",
     "@nomicfoundation/hardhat-ethers": "^3.0.0",
